+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v26.1.5-mc26.1.x] - 2026-05-03
+
+### Added
+
+- Add an additional `ItemComponentsContext::registerItemComponentsPatch` overload
+
+### Fixed
+
+- Fix crash with the Do a Barrel Roll mod
+
 ## [v26.1.4-mc26.1.x] - 2026-04-30
 
 ### Added